Change manual trigger logic
The setting whether a job should be triggered manually is now on the job
itself and not implicitly handled via 'next-manual-stages'. Each job that
should be triggered manually needs to have this setting (as opposed to
the previous logic that all jobs in a manual stage are being triggered
manually).

{
    "stages": [
        {
            "name": "stage0",
            "jobs": [
                { "job0": "echo 'hi' }
            ],
            "next-manual-stages": [ "stage1" ]
        },
        {
            "name": "stage1",
            "jobs": [
                { "job1": "echo 'world'" }
            ]
        }
   ]
   ...
}

needs to be changed to:

{
    "stages": [
        {
            "name": "stage0",
            "jobs": [
                { "job0": "echo 'hi' }
            ],
            "next-stages": [ "stage1" ]
        },
        {
            "name": "stage1",
            "jobs": [
                { "job1": {"cmd": "echo 'world'", "manual": true} }
            ]
        }
   ]
   ...
}